# Optimize DateOnly properties and deconstruct