--- a/src/libraries/System.Private.CoreLib/src/System/Globalization/Calendar.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Globalization/Calendar.cs
@@ -41,19 +41,7 @@ namespace System.Globalization
         internal const int MillisPerHour = MillisPerMinute * 60;
         internal const int MillisPerDay = MillisPerHour * 24;
 
-        // Number of days in a non-leap year
-        internal const int DaysPerYear = 365;
-        // Number of days in 4 years
-        internal const int DaysPer4Years = DaysPerYear * 4 + 1;
-        // Number of days in 100 years
-        internal const int DaysPer100Years = DaysPer4Years * 25 - 1;
-        // Number of days in 400 years
-        internal const int DaysPer400Years = DaysPer100Years * 4 + 1;
-
-        // Number of days from 1/1/0001 to 1/1/10000
-        internal const int DaysTo10000 = DaysPer400Years * 25 - 366;
-
-        internal const long MaxMillis = (long)DaysTo10000 * MillisPerDay;
+        internal const long MaxMillis = (long)DateTime.DaysTo10000 * MillisPerDay;
 
         private int _currentEraValue = -1;