@tarekgh 2024-06-12
I appreciate trying to optimize here but I have a concern the new code is more involving and will be more difficult to maintain. Is there any issue forcing to try optimizing such cases? I mean anyone complaint about the perf of the involved cases.