@lilinus 2024-06-12
> I appreciate trying to optimize here but I have a concern the new code is more involving and will be more difficult to maintain. Is there any issue forcing to try optimizing such cases? I mean anyone complaint about the perf of the cases involved or this optimization affecting any hot path scenario. 

Fair enough.

Working with financial applications; discrete dates (and numerics) is a big part of the values the calculations are done on, and where a significant CPU time is spent.

Thanks!